t/reading-files/lectura-de-archivos closes #37
