+++ public/codificacion-base64
@@ -190,7 +190,7 @@ como se desea.</p>
       
       
       <p class="next">
-        Siguiente ejemplo: <a href="reading-files">Reading Files</a>.
+        Siguiente ejemplo: <a href="lectura-de-archivos">Lectura de archivos</a>.
       </p>
       
       <p class="footer">
