+++ public/index.html
@@ -145,7 +145,7 @@
       
         <li><a href="codificacion-base64">Codificación Base64</a></li>
       
-        <li><a href="reading-files">Reading Files</a></li>
+        <li><a href="lectura-de-archivos">Lectura de archivos</a></li>
       
         <li><a href="escritura-de-archivos">Escritura de archivos</a></li>
